# docs/api/openapi3.json
{
  "openapi": "3.0.0",
  "components": {
    "schemas": {
...
      "DeviceFlowResponse": {
        "properties": {
          "deviceCode": {
            "description": "a code that a device will use to exchange for an access key after device login is approved",
            "example": "NGU4QWFiNjQ5YmQwNG3YTdmZMEyNzQ3YzQ1YSA",
            "type": "string"
          },
          "expiresIn": {
            "description": "The number of seconds that this set of values is valid",
            "example": "1800",
            "format": "int16",
            "type": "integer"
          },
          "interval": {
            "description": "the number of seconds the device should wait between polling to see if the user has finished logging in",
            "example": "5",
            "format": "int8",
            "type": "integer"
          },
          "userCode": {
            "description": "This is the text the user will enter at the Verification URI",
            "example": "BDSD-HQMK",
            "type": "string"
          },
          "verificationURI": {
            "description": "This is the URL the user needs to enter into their browser to start logging in",
            "example": "https://infrahq.com/device",
            "type": "string"
          }
        }
      },
...
      "EmptyResponse": {},
      "Error": {
        "properties": {
          "code": {
            "format": "int32",
            "type": "integer"
          },
          "fieldErrors": {
            "items": {
              "properties": {
                "errors": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                "fieldName": {
                  "type": "string"
                }
              },
              "type": "object"
            },
            "type": "array"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "info": {
    "description": "Infra API",
    "license": {
      "name": "Elastic License v2.0",
      "url": "https://www.elastic.co/licensing/elastic-license"
    },
    "title": "Infra API",
    "version": "0.0.0"
  },
  "paths": {
...
    "/api/access-keys/{id}": {
      "delete": {
        "description": "DeleteAccessKey",
        "operationId": "DeleteAccessKey",
        "parameters": [
          {
            "in": "header",
            "name": "Infra-Version",
            "required": true,
            "schema": {
              "description": "Version of the API being requested",
              "example": "0.0.0",
              "format": "\\d+\\.\\d+\\(.\\d+)?(-.\\w(+\\w)?)?",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "schema": {
              "description": "Bearer followed by your access key",
              "example": "Bearer ACCESSKEY",
              "format": "Bearer [\\da-zA-Z]{10}\\.[\\da-zA-Z]{24}",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "example": "4yJ3n3D8E2",
              "format": "uid",
              "pattern": "[1-9a-km-zA-HJ-NP-Z]{1,11}",
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unauthorized: Requestor is not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Forbidden: Requestor does not have the right permissions"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Duplicate Record"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponse"
                }
              }
            },
            "description": "Success"
          }
        },
        "summary": "DeleteAccessKey",
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/api/device": {
      "post": {
        "description": "StartDeviceFlow",
        "operationId": "StartDeviceFlow",
        "parameters": [
          {
            "in": "header",
            "name": "Infra-Version",
            "required": true,
            "schema": {
              "description": "Version of the API being requested",
              "example": "0.0.0",
              "format": "\\d+\\.\\d+\\(.\\d+)?(-.\\w(+\\w)?)?",
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unauthorized: Requestor is not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Forbidden: Requestor does not have the right permissions"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Duplicate Record"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceFlowResponse"
                }
              }
            },
            "description": "Success"
          }
        },
        "summary": "StartDeviceFlow",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/device/approve": {
      "post": {
        "description": "ApproveDeviceFlow",
        "operationId": "ApproveDeviceFlow",
        "parameters": [
          {
            "in": "header",
            "name": "Infra-Version",
            "required": true,
            "schema": {
              "description": "Version of the API being requested",
              "example": "0.0.0",
              "format": "\\d+\\.\\d+\\(.\\d+)?(-.\\w(+\\w)?)?",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "schema": {
              "description": "Bearer followed by your access key",
              "example": "Bearer ACCESSKEY",
              "format": "Bearer [\\da-zA-Z]{10}\\.[\\da-zA-Z]{24}",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "userCode": {
                    "example": "BDSD-HQMK",
                    "format": "[B-DF-HJ-NP-TV-XZ\\-]",
                    "maxLength": 9,
                    "minLength": 8,
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unauthorized: Requestor is not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Forbidden: Requestor does not have the right permissions"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Duplicate Record"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponse"
                }
              }
            },
            "description": "Success"
          }
        },
        "summary": "ApproveDeviceFlow",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/device/status": {
      "post": {
        "description": "GetDeviceFlowStatus",
        "operationId": "GetDeviceFlowStatus",
        "parameters": [
          {
            "in": "header",
            "name": "Infra-Version",
            "required": true,
            "schema": {
              "description": "Version of the API being requested",
              "example": "0.0.0",
              "format": "\\d+\\.\\d+\\(.\\d+)?(-.\\w(+\\w)?)?",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "deviceCode": {
                    "format": "[a-zA-Z0-9]",
                    "maxLength": 38,
                    "minLength": 38,
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unauthorized: Requestor is not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Forbidden: Requestor does not have the right permissions"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Duplicate Record"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceFlowStatusResponse"
                }
              }
            },
            "description": "Success"
          }
        },
        "summary": "GetDeviceFlowStatus",
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/api/logout": {
      "post": {
        "description": "Logout",
        "operationId": "Logout",
        "parameters": [
          {
            "in": "header",
            "name": "Infra-Version",
            "required": true,
            "schema": {
              "description": "Version of the API being requested",
              "example": "0.0.0",
              "format": "\\d+\\.\\d+\\(.\\d+)?(-.\\w(+\\w)?)?",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "schema": {
              "description": "Bearer followed by your access key",
              "example": "Bearer ACCESSKEY",
              "format": "Bearer [\\da-zA-Z]{10}\\.[\\da-zA-Z]{24}",
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unauthorized: Requestor is not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Forbidden: Requestor does not have the right permissions"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Duplicate Record"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponse"
                }
              }
            },
            "description": "Success"
          }
        },
        "summary": "Logout",
        "tags": [
          "Authentication"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "https://api.infrahq.com"
    }
  ]
}